gcp/requirements.txt
tensorflow==2.11.0
google-cloud-storage==2.7.0
Pillow==9.2.0